requirements.txt
Flask
neo4j
python-dotenv
transformers
torch
PyMuPDF
spacy
PyPDF2
Pillow
pytesseract
anthropic
openai